Enhanced Slurm scheduler with new options
It is now possible to specify the partition to use (by default this setting remains empty, SLURM using the default partition as configured on the system).
It is also possible to set the  option which prevents other jobs from beeing allocated on a host even if there remains unused cores.